article的查找

+++ Shangbao01/src/com/shangbao/dao/ArticleDao.java
@@ -62,7 +62,7 @@ public interface ArticleDao {
      * @param id 
      * @return 
      */  
-    Article findById(int id);  
+    Article findById(long id);  
       
     /** 
      * 查询全部
